requirements.txt
aiofile~=3.9.0
discord.py~=2.6.4